pyproject.toml
[project]
name = "pharma-intelligence"
version = "0.1.0"
description = "Multi-pharmacy price comparison platform for pharmaceutical products"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawlee[all]>=0.4.0",
    "playwright>=1.40.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]
bigquery = [
    "google-cloud-bigquery>=3.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C90"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
